// backend/src/features/meetings/enums.ts
export enum MeetingTasks {
	CHAIRMAN = 'Presidente',
	TREASURES_SPEAKER = 'Tesouros da Palavra e Deus',
	SPIRITUAL_GEMS_DIRECTOR = 'Encontre de Joias Espirituais',
	BOOK_STUDY_DIRECTOR = 'Estudo de Livro',
	BOOK_STUDY_READER = 'Leitor do Estudo de Livro',
	CHRISTIAN_LIFE = 'Nossa Vida Cristã',
}

export enum MeetingFieldMinistry {
	TALK = 'Discurso',
	DEMONSTRATION = 'Apresentação',
	BIBLE_READ = 'Leitura da Bíblia',
}